requirements.txt
# Core dependencies
requests>=2.25.0
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.21.0

# Document processing
beautifulsoup4>=4.9.0
lxml>=4.6.0
html5lib>=1.1

# Vector database and embeddings
chromadb>=0.4.0
sentence-transformers>=2.0.0
google-generativeai>=0.3.0

# Text processing
nltk>=3.6.0

# Async and HTTP
aiohttp>=3.8.0
httpx>=0.24.0

# Utilities
tqdm>=4.60.0
python-dateutil>=2.8.0
pydantic>=1.8.0

# Testing
pytest>=6.0.0
pytest-asyncio>=0.18.0

# Development
jupyter>=1.0.0